Added Plant model
