+++ src/app/services/productservice.service.ts
@@ -1,4 +1,6 @@
 import { Injectable } from '@angular/core';
+import { HttpClient,HttpHeaders} from '@angular/common/http';
+import { Observable } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
