--- src/models/plant.module.ts
@@ -1,10 +1,0 @@
-export interface Plant {
-  id : number,
-  product_name: string,
-  product_description: string,
-  product_price:number,
-  product_image:string,
-  createdAt: Date,
-  updatedAt: Date,
-  userUid: string
- }
